requirements.txt
aiogram>=3.7,<4.0
aiosqlite>=0.20,<1.0
python-dotenv>=1.0,<2.0
apscheduler>=3.10,<4.0
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
vosk>=0.3.45,<0.4
